pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mission-toolkit"
version = "0.1.0"
description = "Mission Design and Flight Sequence Toolkit for small orbital launch vehicles"
requires-python = ">=3.11"
dependencies = [
    "numpy>=1.26",
    "scipy>=1.12",
    "pydantic>=2.6",
    "pyyaml>=6.0",
    "pandas>=2.2",
    "matplotlib>=3.8",
    "rich>=13.7",
    "joblib>=1.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.1",
    "ruff>=0.3",
    "mypy>=1.9",
]
api = [
    "fastapi>=0.115",
    "uvicorn>=0.30",
]

[project.scripts]
run-nominal     = "scripts.run_nominal:main"
run-montecarlo  = "scripts.run_montecarlo:main"
validate-config = "scripts.validate_config:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts   = "--cov=src --cov-report=term-missing -v"

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "UP"]

[tool.mypy]
strict = true
ignore_missing_imports = true
